Fixes #71 - Allow delegation to be bound to intermediate elements between the target and component element similar to as if you were using event delegation with jQuery.
I'm still not sure about the argument order. The e.currentTarget property always refers to the component element making the element parameter unnecessary. The e.target property always refers to the target that fired the event and then we must pass the current element if triggering on an intermediate element. This should be reviewed and possibly changed.