test(vdom): Get single test reproducing erroneous behaviour when going from true to false on a boole
The named-slot polyfill was added because in this test, we are composing two components together.
Without a shadow DOM polyfill the parent component would overwrite the changes the child component
makes to itself if native shadow DOM is not available. This test is only reproducible when two
components are composed together.

#649