feat: Move rendering into the connectedCallback() so that no unnecessary renders occur.
This had some knock on effects to the API that we can no longer depend on the shadow root being
there upon creation of the element. It will be there when it's in the document, but not before. This
makes the ready() callback redundant as well.

BREAKING CHANGE: removed ready() callback. BREAKING CHANGE: removed the render() option from props
in favour of the shouldRender() main callback on the constructor.