feat: Add a renderer() callback to customise the renderer.
This allows a child component to inherit the render() and renderer() from a parent component so that

multiple versions of Incremental DOM aren't brought in (only the one from the parent). While this

isn't bullet proof, it's a good boyscout refactor to contain things more closely to where they're

used in the base Componnet class and covers the major use case of inheriting from a parent where

you're not redefining the renderer or render methods. The added benefit is that renderer() can now

be overridden, which is desirable as we move towards a more decoupled rendering pipeline (even if we

continue to support Incremental DOM).

fixes #845