fix: Remove redundant code that was kept around.
The debounce() utility was updated in Skate 3.0.0 to remove the old setTimeout() code. That was kept
around, but should have been removed. As such, this is a breaking fix and will bump the major
version, but shouldn't affect most people (and should have been broken in 3.0.0, too).

BREAKING CHANGE: Need to supply a MutationObserver polyfill for IE9 / 10.