refactor: Remove public shadowRoot symbol and operation without shadow DOM.
Removed support for symbols.shadowRoot. Consumers can just use the shadowRoot property on the
element. In the future, if the mode is configurable, then attachShadow() can be overridden in
consumerland tests to provide a shadow root for testing purposes. Also removed rendering to the
element if shadowDOM isn't available. To workaround this, attachShadow() can be overridden to return
the element itself.

BREAKING CHANGE: Removed the shadowRoot symbol and shadow DOM v1 is now required for Skate to
function.