fix: Fix proces.env issue with builds.
Removed process.env.NODE_ENV checks as they're kinda long to dev with and just used a DEBUG global
instead. Added replacements to the proper builds in order to remove this in prod mode and ensured
the karma config is up to date. Removed the is-production util and made the deprecated() util a noop
if DEBUG is off.

fixes #976