Fixed multiple definition of properties on the same constructor.
Properties were defined at each instantiation of a given element. These caused the backing symbol to be changed for each instance but as the properties are defined on the prototype, this meant that all instances apart from the last actually lost track of the props value.

This was visible in usage when instantiating the same component multiple times in a page at the same time (with innerHTML for example). Only the last one actually got access to the props when render time comes (after debounce).