Doc changes (#1230)
Various documentation files have been re-written to demonstrate how Skate 5 can be used. 

The README now sports a linear progression from a simple web component with Skate, through creation of mixins, and to using renderers from other libraries like Preact. This in essence serves the purpose of a beginner's tutorial to using Skate, and as such may be extracted to the site contents at a later date.

The CONTRIBUTING guide was updated to include more accurate technical overview of Skate 5.

The example code on the site is now tested and loaded directly from the sample under test.

Some odd things that we'll fix in the future:

* The SSR rendering of the site is stripping the values of the hyperlinks to the renderer repositories.
* The HTML escaping of HTML examples is double-encoding the output.