fix(content): properly return the full content of a Node
`appendChild` removes the child from the parent and isn't pure. As a result when
using it during iteration it will cause nodes to be missed. Instead of using
`appendChild` to build the documentFragment's `childNodes` call `push` directly
upon the array. This prevents weird iteration problems and as far as I'm aware
most closely mirrors how `.content` works in native as it doesn't dismantle the
content of the node and simply returns references to the content in a
documentFragment.